requirements.txt
cffi==1.17.1
cryptography==44.0.0
et_xmlfile==2.0.0
greenlet==3.1.1
msoffcrypto-tool==5.4.2
numpy==2.2.2
olefile==0.47
openpyxl==3.1.5
pandas==2.2.3
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.37
tk==0.1.0
typing_extensions==4.12.2
tzdata==2025.1
